# api/src/main/java/fr/grozeille/db4all/api/dto/ConnectionTestResponse.java
package fr.grozeille.db4all.api.dto;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@Builder
@NoArgsConstructor
@AllArgsConstructor
public class ConnectionTestResponse {
    private boolean success;
    private String message;
}