# api/src/main/resources/static/openapi/datasources-draft.openapi.yaml
openapi: 3.1.0
info:
  title: DB4ALL Datasources Draft API
  version: v0-draft
  description: |
    Draft OpenAPI specification generated from the functional specifications.
    It documents the intended datasource, table mapping, cache and query endpoints,
    even when the corresponding backend endpoints are not implemented yet.
    Current MVP target is a standalone deployment with H2 as the internal application database,
    and LOCAL_FILESYSTEM as the only user-facing datasource type.
    Current MVP authorization is simplified: project administrators can perform all project operations.
servers:
  - url: /
security:
  - bearerAuth: []
  - passwordFlow: []
...
paths:
  /api/v2/projects/{projectId}/datasources:
    get:
      tags:
        - Data Sources
      summary: List project datasources
      description: Returns all datasources configured for a project.
      operationId: listProjectDatasources
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      responses:
        '200':
          description: Datasources retrieved.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DatasourceSummary'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/ProjectNotFound'
    post:
      tags:
        - Data Sources
      summary: Create project datasource
      description: |
        Creates a datasource for a project and optionally validates connectivity before saving.
        In the current MVP target, only LOCAL_FILESYSTEM is implemented.
      operationId: createProjectDatasource
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DatasourceCreateRequest'
      responses:
        '201':
          description: Datasource created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DatasourceDetail'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/ProjectNotFound'
  /api/v2/projects/{projectId}/datasources/{datasourceId}:
    get:
      tags:
        - Data Sources
      summary: Get datasource details
      operationId: getProjectDatasource
      parameters:
        - $ref: '#/components/parameters/ProjectId'
        - $ref: '#/components/parameters/DatasourceId'
      responses:
        '200':
          description: Datasource retrieved.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DatasourceDetail'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/DatasourceNotFound'
    put:
      tags:
        - Data Sources
      summary: Update datasource
      operationId: updateProjectDatasource
      parameters:
        - $ref: '#/components/parameters/ProjectId'
        - $ref: '#/components/parameters/DatasourceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DatasourceUpdateRequest'
      responses:
        '200':
          description: Datasource updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DatasourceDetail'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/DatasourceNotFound'
    delete:
      tags:
        - Data Sources
      summary: Delete datasource
      operationId: deleteProjectDatasource
      parameters:
        - $ref: '#/components/parameters/ProjectId'
        - $ref: '#/components/parameters/DatasourceId'
      responses:
        '204':
          description: Datasource deleted.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/DatasourceNotFound'
  /api/v2/projects/{projectId}/datasources/{datasourceId}/test-connection:
    post:
      tags:
        - Data Sources
      summary: Test datasource connection
      description: Verifies that the datasource configuration can be used successfully.
      operationId: testDatasourceConnection
      parameters:
        - $ref: '#/components/parameters/ProjectId'
        - $ref: '#/components/parameters/DatasourceId'
      responses:
        '200':
          description: Connection successful.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConnectionTestResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/DatasourceNotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    passwordFlow:
      type: oauth2
      flows:
        password:
          tokenUrl: /api/v2/auth/login
  parameters:
    ProjectId:
      name: projectId
      in: path
      required: true
      schema:
        type: string
    DatasourceId:
      name: datasourceId
      in: path
      required: true
      schema:
        type: string
...
  responses:
    Unauthorized:
      description: Authentication required.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Access denied.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ValidationError:
      description: Invalid request.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    ProjectNotFound:
      description: Project not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    DatasourceNotFound:
      description: Datasource not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
    ErrorResponse:
      type: object
      required:
        - message
      properties:
        message:
          type: string
        code:
          type: string
    DatasourceType:
      type: string
      description: In the current MVP target, only LOCAL_FILESYSTEM is implemented.
      enum:
        - LOCAL_FILESYSTEM
        - S3
        - AZURE_STORAGE
        - POSTGRESQL
        - CIFS
    DatasourceSummary:
      type: object
      required:
        - id
        - name
        - type
        - readOnly
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        type:
          $ref: '#/components/schemas/DatasourceType'
        readOnly:
          type: boolean
        cacheDefaultPolicySummary:
          type: string
    DatasourceDetail:
      allOf:
        - $ref: '#/components/schemas/DatasourceSummary'
        - type: object
          required:
            - configuration
          properties:
            configuration:
              $ref: '#/components/schemas/DatasourceConfiguration'
    DatasourceCreateRequest:
      allOf:
        - $ref: '#/components/schemas/DatasourceUpsertBase'
        - type: object
          required:
            - name
    DatasourceUpdateRequest:
      allOf:
        - $ref: '#/components/schemas/DatasourceUpsertBase'
    DatasourceUpsertBase:
      type: object
      required:
        - type
        - configuration
      properties:
        name:
          type: string
        description:
          type: string
        readOnly:
          type: boolean
          default: true
        type:
          $ref: '#/components/schemas/DatasourceType'
        configuration:
          $ref: '#/components/schemas/DatasourceConfiguration'
    DatasourceConfiguration:
      oneOf:
        - $ref: '#/components/schemas/LocalFilesystemDatasourceConfiguration'
        - $ref: '#/components/schemas/S3DatasourceConfiguration'
        - $ref: '#/components/schemas/AzureStorageDatasourceConfiguration'
        - $ref: '#/components/schemas/PostgreSqlDatasourceConfiguration'
        - $ref: '#/components/schemas/CifsDatasourceConfiguration'
      discriminator:
        propertyName: type
        mapping:
          LOCAL_FILESYSTEM: '#/components/schemas/LocalFilesystemDatasourceConfiguration'
          S3: '#/components/schemas/S3DatasourceConfiguration'
          AZURE_STORAGE: '#/components/schemas/AzureStorageDatasourceConfiguration'
          POSTGRESQL: '#/components/schemas/PostgreSqlDatasourceConfiguration'
          CIFS: '#/components/schemas/CifsDatasourceConfiguration'
    LocalFilesystemDatasourceConfiguration:
      type: object
      description: Current MVP datasource type for standalone demos and local usage.
      required:
        - type
        - rootPath
      properties:
        type:
          type: string
          const: LOCAL_FILESYSTEM
        rootPath:
          type: string
    S3DatasourceConfiguration:
      type: object
      required:
        - type
        - host
        - bucket
        - accessKey
        - secretKey
      properties:
        type:
          type: string
          const: S3
        host:
          type: string
        port:
          type: integer
        bucket:
          type: string
        basePathPrefix:
          type: string
        accessKey:
          type: string
        secretKey:
          type: string
          format: password
        region:
          type: string
        advancedOptions:
          type: object
          additionalProperties:
            type: string
    AzureStorageDatasourceConfiguration:
      type: object
      required:
        - type
        - accountName
        - container
        - credentialType
      properties:
        type:
          type: string
          const: AZURE_STORAGE
        accountName:
          type: string
        container:
          type: string
        basePathPrefix:
          type: string
        credentialType:
          type: string
          enum:
            - SHARED_KEY
            - SAS_TOKEN
        sharedKey:
          type: string
          format: password
        sasToken:
          type: string
          format: password
    PostgreSqlDatasourceConfiguration:
      type: object
      required:
        - type
        - hostname
        - port
        - databaseName
        - login
        - password
      properties:
        type:
          type: string
          const: POSTGRESQL
        hostname:
          type: string
        port:
          type: integer
        databaseName:
          type: string
        schemaFilter:
          type: string
        login:
          type: string
        password:
          type: string
          format: password
        connectionProperties:
          type: object
          additionalProperties:
            type: string
    CifsDatasourceConfiguration:
      type: object
      required:
        - type
        - path
        - login
        - password
      properties:
        type:
          type: string
          const: CIFS
        path:
          type: string
          description: UNC path or server/share path.
        login:
          type: string
        password:
          type: string
          format: password
        syncFrequency:
          $ref: '#/components/schemas/SyncFrequency'
        usageNotes:
          type: string
          example: Use only for small files. Prefer S3 or Azure Storage for large CSV or Parquet datasets.
    SyncFrequency:
      type: object
      required:
        - value
        - unit
      properties:
        value:
          type: integer
          minimum: 1
        unit:
          type: string
          enum:
            - HOURS
            - DAYS
        cron:
          type: string
          description: Optional advanced scheduling expression.
    ConnectionTestResponse:
      type: object
      required:
        - success
      properties:
        success:
          type: boolean
        message:
          type: string
...